+++ simple_history/models.py
@@ -2,40 +2,33 @@ from __future__ import unicode_literals
 
 import threading
 import copy
-from django.db.models import Q
-
-try:
-    from django.apps import apps  # Django >= 1.7
-except ImportError:
-    apps = None
+import warnings
 from django.db import models, router
-from django.db.models import loading
+from django.db.models import loading, Q
 from django.db.models.fields.proxy import OrderWrt
 from django.db.models.fields.related import RelatedField
-from django.db.models.related import RelatedObject
 from django.conf import settings
 from django.contrib import admin
 from django.utils import importlib, six
 from django.utils.encoding import python_2_unicode_compatible
-try:
-    from django.utils.encoding import smart_text
-except ImportError:
-    from django.utils.encoding import smart_unicode as smart_text
-try:
-    from django.utils.timezone import now
-except ImportError:
-    from datetime import datetime
-    now = datetime.now
+from django.utils.encoding import smart_text
+from django.utils.timezone import now
 from django.utils.translation import string_concat
-try:
-    from south.modelsinspector import add_introspection_rules
-except ImportError:
-    pass
-else:
-    add_introspection_rules(
-        [], ["^simple_history.models.CustomForeignKeyField"])
+
 from simple_history import register
 from .manager import HistoryDescriptor
+
+try:
+    from django.apps import apps
+except ImportError:  # Django < 1.7
+    apps = None
+try:
+    from south.modelsinspector import add_introspection_rules
+except ImportError:  # south not present
+    pass
+else:  # south configuration for CustomForeignKeyField
+    add_introspection_rules(
+        [], ["^simple_history.models.CustomForeignKeyField"])
 
 ALL_M2M_FIELDS = object()
 
@@ -51,7 +44,7 @@ def not_registered(model):
 class HistoricalRecords(object):
     thread = threading.local()
 
-    def __init__(self, verbose_name=None, bases=(models.Model,), user_related_name=None, m2m_fields=None):
+    def __init__(self, verbose_name=None, bases=(models.Model,), user_related_name='+', m2m_fields=None):
         self.user_set_verbose_name = verbose_name
         self.user_related_name = user_related_name
         self.m2m_fields = m2m_fields
@@ -166,7 +159,7 @@ class HistoricalRecords(object):
             # registered under different app
             attrs['__module__'] = self.module
         elif app_module != self.module:
-            if apps is None:
+            if apps is None:  # Django < 1.7
                 # has meta options with app_label
                 app = loading.get_app(model._meta.app_label)
                 attrs['__module__'] = app.__name__  # full dotted name
@@ -194,18 +187,25 @@ class HistoricalRecords(object):
         for field in model._meta.fields:
             field = copy.copy(field)
             field.rel = copy.copy(field.rel)
-            if isinstance(field, models.ForeignKey):
-                # Don't allow reverse relations.
-                # ForeignKey knows best what datatype to use for the column
-                # we'll used that as soon as it's finalized by copying rel.to
-                field.__class__ = CustomForeignKeyField
-                field.rel.related_name = '+'
-                field.null = True
-                field.blank = True
             if isinstance(field, OrderWrt):
                 # OrderWrt is a proxy field, switch to a plain IntegerField
                 field.__class__ = models.IntegerField
-            transform_field(field)
+            if isinstance(field, models.ForeignKey):
+                old_field = field
+                field = type(field)(
+                    field.rel.to,
+                    related_name='+',
+                    null=True,
+                    blank=True,
+                    primary_key=False,
+                    db_index=True,
+                    serialize=True,
+                )
+                field._unique = False
+                field.name = old_field.name
+                field.db_constraint = False
+            else:
+                transform_field(field)
             fields[field.name] = field
         return fields
 
@@ -220,7 +220,7 @@ class HistoricalRecords(object):
             opts = model._meta
             try:
                 app_label, model_name = opts.app_label, opts.model_name
-            except AttributeError:
+            except AttributeError:  # Django < 1.7
                 app_label, model_name = opts.app_label, opts.module_name
             return ('%s:%s_%s_simple_history' %
                     (admin.site.name, app_label, model_name),
@@ -233,7 +233,8 @@ class HistoricalRecords(object):
             'history_id': models.AutoField(primary_key=True),
             'history_date': models.DateTimeField(),
             'history_user': models.ForeignKey(
-                user_model, null=True, related_name=self.user_related_name),
+                user_model, null=True, related_name=self.user_related_name,
+                on_delete=models.SET_NULL),
             'history_type': models.CharField(max_length=1, choices=(
                 ('+', 'Created'),
                 ('~', 'Changed'),
@@ -254,6 +255,7 @@ class HistoricalRecords(object):
         """
         meta_fields = {
             'ordering': ('-history_date', '-history_id'),
+            'get_latest_by': 'history_date',
         }
         if self.user_set_verbose_name:
             name = self.user_set_verbose_name
@@ -348,6 +350,8 @@ class HistoricalRecords(object):
 class CustomForeignKeyField(models.ForeignKey):
 
     def __init__(self, *args, **kwargs):
+        warnings.warn("CustomForeignKeyField is deprecated.",
+                      DeprecationWarning)
         super(CustomForeignKeyField, self).__init__(*args, **kwargs)
         self.db_constraint = False
         self.generate_reverse_relation = False
@@ -410,13 +414,6 @@ class CustomForeignKeyField(models.ForeignKey):
 
     def do_related_class(self, other, cls):
         field = self.get_field(other, cls)
-        if not hasattr(self, 'related'):
-            try:
-                instance_type = cls.instance_type
-            except AttributeError:  # when model is reconstituted for migration
-                pass  # happens during migrations
-            else:
-                self.related = RelatedObject(other, instance_type, self)
         transform_field(field)
         field.rel = None