+++ simple_history/tests/models.py
@@ -232,3 +232,16 @@ class SeriesWork(models.Model):
 
     class Meta:
         order_with_respect_to = 'series'
+
+
+class PollInfo(models.Model):
+    poll = models.ForeignKey(Poll, primary_key=True)
+    history = HistoricalRecords()
+
+
+class UserAccessorDefault(models.Model):
+    pass
+
+
+class UserAccessorOverride(models.Model):
+    pass