+++ simple_history/tests/tests/test_admin.py
@@ -1,7 +1,6 @@
 from datetime import datetime, timedelta
 from django_webtest import WebTest
 from django.test.utils import override_settings
-from django.test import TestCase
 from django import VERSION
 from django.core.urlresolvers import reverse
 try:
@@ -12,7 +11,6 @@ except ImportError:  # django 1.4 compatibility
 from django.contrib.admin.util import quote
 from django.conf import settings
 from simple_history.models import HistoricalRecords
-from simple_history.templatetags import simple_history_compare
 
 from ..models import Book, Person, Poll, State
 
@@ -226,39 +224,17 @@ class AdminSiteTest(AdminTest):
         change_url = get_history_url(state, 0, site="other_admin")
         self.app.get(change_url)
 
-    def test_compare_history(self):
+    def test_deleteting_user(self):
+        """Test deletes of a user does not cascade delete the history"""
         self.login()
+        poll = Poll(question="why?", pub_date=today)
+        poll._history_user = self.user
+        poll.save()
 
+        historical_poll = poll.history.all()[0]
+        self.assertEqual(historical_poll.history_user, self.user)
 
-class CompareHistoryTest(AdminTest):
+        self.user.delete()
 
-    def setUp(self):
-        super(CompareHistoryTest, self).setUp()
-        self.login()
-        self.poll = Poll.objects.create(question="Who?", pub_date=today)
-        for question in ("What?", "Where?", "When?", "Why?", "How?"):
-            self.poll.question = question
-            self.poll.save()
-
-    def test_navigate_to_compare(self):
-        response = self.app.get(get_history_url(self.poll)).form.submit()
-        response.mustcontain("<title>Compare ")
-
-
-class CompareTableTest(TestCase):
-
-    def test_diff_table(self):
-        table_markup = simple_history_compare.diff_table(a="this\nan\ntest", b="this\nis\na\ntest")
-        self.assertEqual(table_markup, """
-    <table class="diff" id="difflib_chg_to1__top"
-           cellspacing="0" cellpadding="0" rules="groups" >
-        <colgroup></colgroup> <colgroup></colgroup> <colgroup></colgroup>
-        <colgroup></colgroup> <colgroup></colgroup> <colgroup></colgroup>
-        
-        <tbody>
-            <tr><td class="diff_next" id="difflib_chg_to1__0"><a href="#difflib_chg_to1__0">f</a></td><td class="diff_header" id="from1_1">1</td><td nowrap="nowrap">this</td><td class="diff_next"><a href="#difflib_chg_to1__0">f</a></td><td class="diff_header" id="to1_1">1</td><td nowrap="nowrap">this</td></tr>
-            <tr><td class="diff_next"><a href="#difflib_chg_to1__top">t</a></td><td class="diff_header" id="from1_2">2</td><td nowrap="nowrap"><span class="diff_sub">an</span></td><td class="diff_next"><a href="#difflib_chg_to1__top">t</a></td><td class="diff_header" id="to1_2">2</td><td nowrap="nowrap"><span class="diff_add">is</span></td></tr>
-            <tr><td class="diff_next"></td><td class="diff_header"></td><td nowrap="nowrap"></td><td class="diff_next"></td><td class="diff_header" id="to1_3">3</td><td nowrap="nowrap"><span class="diff_add">a</span></td></tr>
-            <tr><td class="diff_next"></td><td class="diff_header" id="from1_3">3</td><td nowrap="nowrap">test</td><td class="diff_next"></td><td class="diff_header" id="to1_4">4</td><td nowrap="nowrap">test</td></tr>
-        </tbody>
-    </table>""")
+        historical_poll = poll.history.all()[0]
+        self.assertEqual(historical_poll.history_user, None)